+++ packages/shared/src/components/index.ts
@@ -18,6 +18,7 @@ export {
   useToggleButtonGroupIndicator,
 } from "./ToggleButtonGroupContext";
 export { TextField } from "./TextField";
+export { TextArea } from "./TextArea";
 export { NumberField } from "./NumberField";
 export { SearchField } from "./SearchField";
 export { Checkbox } from "./Checkbox";